requirements.txt
google-cloud-aiplatform
streamlit
pydub
google-cloud-texttospeech
google-cloud-vision
google-cloud-storage
ffmpeg-python
fastapi
uvicorn
requests
gunicorn
python-multipart
python-dotenv
celery[redis]
redis
flower
pyyaml
